Refactoring of pilot, created new arguments module
