+++ pilot.py
@@ -23,7 +23,6 @@
 
 """This is the entry point for the PanDA Pilot, executed with 'python3 pilot.py <args>'."""
 
-import argparse
 import logging
 import os
 import sys
@@ -31,10 +30,10 @@ import threading
 import time
 from os import getcwd, chdir, environ
 from os.path import exists, join
-from re import match
 from shutil import rmtree
 from typing import Any
 
+from arguments import get_args
 from pilot.common.errorcodes import ErrorCodes
 from pilot.common.exception import PilotException
 from pilot.common.pilotcache import get_pilot_cache
@@ -262,489 +261,6 @@ def check_cvmfs(logger: Any) -> int:
         return errors.CVMFSISNOTALIVE
 
     return 0
-
-
-def str2bool(var: str) -> bool:
-    """
-    Convert string to bool.
-
-    Args:
-        var (str): String to be converted to bool.
-
-    Returns:
-        bool: Converted boolean value.
-    """
-    if isinstance(var, bool):  # does this ever happen?
-        return var
-
-    if var.lower() in {"yes", "true", "t", "y", "1"}:
-        ret = True
-    elif var.lower() in {"no", "false", "f", "n", "0"}:
-        ret = False
-    else:
-        raise argparse.ArgumentTypeError(f"boolean value expected (var={var})")
-
-    return ret
-
-
-def validate_resource_type(value: str) -> str:
-    """
-    Validate the resource type.
-
-    Args:
-        value (str): Resource type.
-
-    Returns:
-        str: Validated resource type.
-
-    Raises:
-        argparse.ArgumentTypeError: If the resource type is invalid.
-    """
-    # Define the allowed patterns
-    allowed_patterns = ["", "SCORE", "MCORE", "SCORE_*", "MCORE_*"]
-    if value in allowed_patterns:
-        return value
-    # Check for pattern matching
-    for pattern in allowed_patterns:
-        if pattern.endswith('*') and match(f"^{pattern[:-1]}", value):
-            return value
-    raise argparse.ArgumentTypeError(f"Invalid resource type: {value}")
-
-
-def get_args() -> Any:
-    """
-    Return the args from the arg parser.
-
-    Returns:
-        argparse.Namespace: Parsed command-line arguments.
-    """
-    arg_parser = argparse.ArgumentParser()
-
-    # pilot log creation
-    arg_parser.add_argument(
-        "--no-pilot-log",
-        dest="nopilotlog",
-        action="store_true",
-        default=False,
-        help="Do not write the pilot log to file",
-    )
-
-    # pilot work directory
-    arg_parser.add_argument(
-        "-a", "--workdir", dest="workdir", default="", help="Pilot work directory"
-    )
-
-    # debug option to enable more log messages
-    arg_parser.add_argument(
-        "-d",
-        "--debug",
-        dest="debug",
-        action="store_true",
-        default=False,
-        help="Enable debug mode for logging messages",
-    )
-
-    # the choices must match in name the python module in pilot/workflow/
-    arg_parser.add_argument(
-        "-w",
-        "--workflow",
-        dest="workflow",
-        default="generic",
-        choices=[
-            "generic",
-            "generic_hpc",
-            "production",
-            "production_hpc",
-            "analysis",
-            "analysis_hpc",
-            "eventservice_hpc",
-            "stager",
-            "payload_stageout",
-        ],
-        help="Pilot workflow (default: generic)",
-    )
-
-    # graciously stop pilot process after hard limit
-    arg_parser.add_argument(
-        "-l",
-        "--lifetime",
-        dest="lifetime",
-        default=324000,
-        required=False,
-        type=int,
-        help="Pilot lifetime seconds (default: 324000 s)",
-    )
-    arg_parser.add_argument(
-        "-L",
-        "--leasetime",
-        dest="leasetime",
-        default=3600,
-        required=False,
-        type=int,
-        help="Pilot leasetime seconds (default: 3600 s)",
-    )
-
-    # Disabe cvmfs checks
-    arg_parser.add_argument(
-        "-b",
-        "--nocvmfs",
-        dest="cvmfs",
-        action="store_false",
-        default=True,
-        help="Disable cvmfs checks",
-    )
-
-    # set the appropriate site, resource and queue
-    arg_parser.add_argument(
-        "-q",
-        "--queue",
-        dest="queue",
-        required=True,
-        help="MANDATORY: queue name (e.g., AGLT2_TEST-condor)",
-    )
-    arg_parser.add_argument(
-        "-r",
-        "--resource",
-        dest="resource",
-        required=False,  # From v 2.2.0 the resource name is internally set
-        help="OBSOLETE: resource name (e.g., AGLT2_TEST)",
-    )
-    arg_parser.add_argument(
-        "-s",
-        "--site",
-        dest="site",
-        required=False,  # From v 2.2.1 the site name is internally set
-        help="OBSOLETE: site name (e.g., AGLT2_TEST)",
-    )
-    arg_parser.add_argument(
-        "-j",
-        "--joblabel",
-        dest="job_label",
-        default="ptest",
-        help="Job prod/source label (default: ptest)",
-    )
-    arg_parser.add_argument(
-        "-g",
-        "--baseurls",
-        dest="baseurls",
-        default="",
-        help="Comma separated list of base URLs for validation of trf download",
-    )
-
-    # pilot version tag; PR or RC
-    arg_parser.add_argument(
-        "-i",
-        "--versiontag",
-        dest="version_tag",
-        default="PR",
-        help="Version tag (default: PR, optional: RC)",
-    )
-
-    arg_parser.add_argument(
-        "-z",
-        "--noserverupdate",
-        dest="update_server",
-        action="store_false",
-        default=True,
-        help="Disable server updates",
-    )
-
-    arg_parser.add_argument(
-        "-k",
-        "--noworkerpilotstatusupdate",
-        dest="workerpilotstatusupdate",
-        action="store_false",
-        default=True,
-        help="Disable updates to updateWorkerPilotStatus",
-    )
-
-    arg_parser.add_argument(
-        "-t",
-        "--noproxyverification",
-        dest="verify_proxy",
-        action="store_false",
-        default=True,
-        help="Disable proxy verification",
-    )
-
-    arg_parser.add_argument(
-        "-u",
-        "--verifypayloadproxy",
-        dest="verify_payload_proxy",
-        action="store_false",
-        default=True,
-        help="Disable payload proxy verification",
-    )
-
-    # graciously stop pilot process after hard limit
-    arg_parser.add_argument(
-        "-v",
-        "--getjobrequests",
-        dest="getjob_requests",
-        default=2,
-        required=False,
-        type=int,
-        help="Number of getjob requests",
-    )
-
-    arg_parser.add_argument(
-        "-x",
-        "--getjobfailures",
-        dest="getjob_failures",
-        default=5,
-        required=False,
-        type=int,
-        help="Maximum number of getjob request failures in Harvester mode",
-    )
-
-    # no_token_renewal
-    arg_parser.add_argument(
-        "-y",
-        "--notokenrenewal",
-        dest="token_renewal",
-        action="store_false",
-        default=True,
-        help="Disable token renewal",
-    )
-
-    arg_parser.add_argument(
-        "--subscribe-to-msgsvc",
-        dest="subscribe_to_msgsvc",
-        action="store_true",
-        default=False,
-        required=False,
-        help="Ask Pilot to receive job/task info from ActiveMQ",
-    )
-
-    # SSL certificates
-    arg_parser.add_argument(
-        "--cacert",
-        dest="cacert",
-        default=None,
-        help="CA certificate to use with HTTPS calls to server, commonly X509 proxy",
-        metavar="path/to/your/certificate",
-    )
-    arg_parser.add_argument(
-        "--capath",
-        dest="capath",
-        default=None,
-        help="CA certificates path",
-        metavar="path/to/certificates/",
-    )
-
-    # Server URLs and ports
-    arg_parser.add_argument(
-        "--url",
-        dest="url",
-        default="",  # the proper default is stored in default.cfg
-        help="PanDA server URL",
-    )
-    arg_parser.add_argument(
-        "-p",
-        "--port",
-        dest="port",
-        type=int,
-        default=25443,
-        help="PanDA server port"
-    )
-    arg_parser.add_argument(
-        "--queuedata-url",
-        dest="queuedata_url",
-        default="",
-        help="Queuedata server URL"
-    )
-    arg_parser.add_argument(
-        "--storagedata-url",
-        dest="storagedata_url",
-        default="",
-        help="URL for downloading DDM end points data",
-    )
-    arg_parser.add_argument(
-        "--rucio-host",
-        dest="rucio_host",
-        default="",
-        help="URL for the Rucio host (optional)",
-    )
-    arg_parser.add_argument(
-        "--redirect-stdout",
-        dest="redirectstdout",
-        default="",
-        help="Redirect all stdout to given file, or /dev/null (optional)",
-    )
-
-    # Country group
-    arg_parser.add_argument(
-        "--country-group",
-        dest="country_group",
-        default="",
-        help="Country group option for getjob request",
-    )
-
-    # Working group
-    arg_parser.add_argument(
-        "--working-group",
-        dest="working_group",
-        default="",
-        help="Working group option for getjob request",
-    )
-
-    # Allow other country
-    arg_parser.add_argument(
-        "--allow-other-country",
-        dest="allow_other_country",
-        type=str2bool,
-        default=False,
-        help="Is the resource allowed to be used outside the privileged group?",
-    )
-
-    # Allow same user
-    arg_parser.add_argument(
-        "--allow-same-user",
-        dest="allow_same_user",
-        type=str2bool,
-        default=True,
-        help="Multi-jobs will only come from same taskID (and thus same user)",
-    )
-
-    # Experiment
-    arg_parser.add_argument(
-        "--pilot-user",
-        dest="pilot_user",
-        default="generic",
-        required=True,
-        help="Pilot user (e.g. name of experiment corresponding to pilot plug-in)",
-    )
-
-    # Kubernetes (pilot running in a pod)
-    arg_parser.add_argument(
-        "--pod",
-        dest="pod",
-        action="store_true",
-        default=False,
-        help="Pilot running in a Kubernetes pod",
-    )
-
-    # Harvester specific options (if any of the following options are used, args.harvester will be set to True)
-    arg_parser.add_argument(
-        "--harvester-workdir",
-        dest="harvester_workdir",
-        default="",
-        help="Harvester work directory",
-    )
-    arg_parser.add_argument(
-        "--harvester-datadir",
-        dest="harvester_datadir",
-        default="",
-        help="Harvester data directory",
-    )
-    arg_parser.add_argument(
-        "--harvester-eventstatusdump",
-        dest="harvester_eventstatusdump",
-        default="",
-        help="Harvester event status dump json file containing processing status",
-    )
-    arg_parser.add_argument(
-        "--harvester-workerattributes",
-        dest="harvester_workerattributes",
-        default="",
-        help="Harvester worker attributes json file containing job status",
-    )
-    arg_parser.add_argument(
-        "--harvester-submit-mode",
-        dest="harvester_submitmode",
-        default="PULL",
-        help="Harvester submit mode (PUSH or PULL [default])",
-    )
-    arg_parser.add_argument(
-        "--resource-type",
-        dest="resource_type",
-        default="",
-        type=validate_resource_type,
-        help="Resource type; MCORE, SCORE or patterns like SCORE_* and MCORE_*",
-    )
-    arg_parser.add_argument(
-        "--use-https",
-        dest="use_https",
-        type=str2bool,
-        default=True,
-        help="Use HTTPS protocol for communications with server",
-    )
-    arg_parser.add_argument(
-        "--cleanup",
-        dest="cleanup",
-        type=str2bool,
-        default=True,
-        help="Cleanup work directory after pilot has finished",
-    )
-    arg_parser.add_argument(
-        "--use-realtime-logging",
-        dest="use_realtime_logging",
-        action="store_true",
-        default=False,
-        help="Use near real-time logging, default=False",
-    )
-    arg_parser.add_argument(
-        "--realtime-logging-server",
-        dest="realtime_logging_server",
-        default=None,
-        help="Realtime logging server [type:host:port]; e.g. logstash:12.23.34.45:80",
-    )
-    # arg_parser.add_argument('--realtime-logfiles',
-    #                         dest='realtime_logfiles',
-    #                         default=None,
-    #                         help='The log filenames to be sent to the realtime logging server')
-    arg_parser.add_argument(
-        "--realtime-logname",
-        dest="realtime_logname",
-        default=None,
-        help="The log name in the realtime logging server",
-    )
-
-    # Harvester and Nordugrid specific options
-    arg_parser.add_argument(
-        "--input-dir", dest="input_dir", default="", help="Input directory"
-    )
-    arg_parser.add_argument(
-        "--input-destination-dir",
-        dest="input_destination_dir",
-        default="",
-        help="Input destination directory",
-    )
-    arg_parser.add_argument(
-        "--output-dir", dest="output_dir", default="", help="Output directory"
-    )
-    arg_parser.add_argument(
-        "--job-type", dest="jobtype", default="", help="Job type (managed, user)"
-    )
-    arg_parser.add_argument(
-        "--use-rucio-traces",
-        dest="use_rucio_traces",
-        type=str2bool,
-        default=True,
-        help="Use rucio traces",
-    )
-
-    # HPC options
-    arg_parser.add_argument(
-        "--hpc-resource",
-        dest="hpc_resource",
-        default="",
-        help="Name of the HPC (e.g. Titan)",
-    )
-    arg_parser.add_argument(
-        "--hpc-mode",
-        dest="hpc_mode",
-        default="manytoone",
-        help="HPC mode (manytoone, jumbojobs)",
-    )
-    arg_parser.add_argument(
-        "--es-executor-type",
-        dest="executor_type",
-        default="generic",
-        help="Event service executor type (generic, raythena)",
-    )
-
-    return arg_parser.parse_args()
 
 
 def create_main_work_dir() -> (int, str):
